Consolidate Interactive Graph tests and stories
