+++ packages/perseus/src/widgets/__testdata__/randomizers.test.ts
@@ -5,7 +5,7 @@ import {
     randomLetter,
     randomSentence,
     randomWord,
-} from "../__testdata__/randomizers";
+} from "./randomizers";
 
 describe("randomizers", () => {
     describe("randomLetter", () => {
