+++ packages/perseus/src/widgets/expression/expression.test.tsx
@@ -8,7 +8,7 @@ import {
 } from "../../../../../testing/test-dependencies";
 import * as Dependencies from "../../dependencies";
 import {mockStrings} from "../../strings";
-import {renderQuestion} from "../__tests__/renderQuestion";
+import {renderQuestion} from "../__testutils__/renderQuestion";
 
 import {Expression} from "./expression";
 import {
