+++ packages/perseus/src/widgets/graded-group-set/graded-group-set-jipt.test.ts
@@ -6,7 +6,7 @@
 
 import {testDependencies} from "../../../../../testing/test-dependencies";
 import * as Dependencies from "../../dependencies";
-import {renderQuestion} from "../__tests__/renderQuestion";
+import {renderQuestion} from "../__testutils__/renderQuestion";
 
 import {article1} from "./graded-group-set.testdata";
