+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph-regression.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 
 import Renderer from "../../renderer";
-import {mockStrings} from "../../strings";
-import {interactiveGraphQuestionBuilder} from "../interactive-graphs/interactive-graph-question-builder";
+import {mockStrings} from "@khanacademy/perseus/strings";
+import {interactiveGraphQuestionBuilder} from "./interactive-graph-question-builder";
 
 import type {PerseusRenderer} from "../../perseus-types";
