+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph.testdata.ts
@@ -1,8 +1,8 @@
-import {interactiveGraphQuestionBuilder} from "../interactive-graphs/interactive-graph-question-builder";
+import {interactiveGraphQuestionBuilder} from "./interactive-graph-question-builder";
 
 import type {Coord} from "../../interactive2/types";
 import type {PerseusRenderer, RadioWidget} from "../../perseus-types";
-import type {LockedFunctionOptions} from "../interactive-graphs/interactive-graph-question-builder";
+import type {LockedFunctionOptions} from "./interactive-graph-question-builder";
 
 // Data for the interactive graph widget
