+++ packages/perseus/src/widgets/passage/__tests__/passage.test.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import {testDependencies} from "../../../../../../testing/test-dependencies";
 import * as Dependencies from "../../../dependencies";
 import {ApiOptions} from "../../../perseus-api";
-import {renderQuestion} from "../../__tests__/renderQuestion";
+import {renderQuestion} from "../../__testutils__/renderQuestion";
 import PassageWidgetExport, {LineHeightMeasurer} from "../passage";
 
 import {question1, question2} from "./passage.testdata";
