+++ packages/perseus/src/widgets/radio/__tests__/radio.test.ts
@@ -5,7 +5,7 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 import {clone} from "../../../../../../testing/object-utils";
 import {testDependencies} from "../../../../../../testing/test-dependencies";
 import * as Dependencies from "../../../dependencies";
-import {renderQuestion} from "../../__tests__/renderQuestion";
+import {renderQuestion} from "../../__testutils__/renderQuestion";
 import PassageWidget from "../../passage";
 
 import {
